logmux - smush many log streams to one logstash
In our app.iced docker environment, we're getting logs from
a bunch of different places and want to output them all to a single
logstash service running locally.

The program: (1) opens a TCP connection to such a logstash; and (2)
mux'es multiple incoming log streams into that TCP connection. It
furthermore adds a "tag" to each logline so we know where it's coming
from on the ELK side.